[panorama] Fix combo score showing wrong values:
For every `ComboScore` input that results in spinning a ticker, it enqueues the spin command in the
ticker's spin queue.

If the ticker is not spinning, it triggers the queue consumption.

The ticker's queue is consumed by taking the earliest spin command from the queue and executing it,
ignoring the older commands.

When the spin command starts, the ticker is flagged as spinning.

If subsequent inputs are made while the ticker is flagged, it will enqueue the following spin
commands.

When the spin command is finished, its spinning flag is removed and it triggers queue consumption
once again. At the end of a series of inputs, if the ticker is spinning, there are
no more incoming inputs to trigger queue consumption after the ticker stops spinning, therefore this
last "forced" trigger is needed.

This assumes the inputs comes in order (the earliest input corresponds to the most up to date
value), which is usually the case for damage accumulation.

Fixes #11.